flats: replace SWATER with fireblu. (#1137)
* flats: replace SWATER with fireblu.

This was briefly discussed on Discord a while ago and I'll throw it out here and see if it sticks.

Basically nobody has any use case for that choppy bright yellow and blue animation. It doesn't really look like anything, so it's either:

1. some hazardous undefined pollutant for a hurtfloor
2. some high-contrast dimensional-rift warp sort of deal

Between nukage, blood, brown sludge and lava - heck, even blue water is good for radioactive stuff - we've pretty much got every use case covered. I've actually never seen the current SWATER used this way.

A lot of messed-up dimensional-rift stuff is represented by a combination of hot rock flats and fireblu - in a lot of places where the hot rock flat is pretty obviously motivated by the lack of a fireblu flat.

This proposed change will allow that fireblu to be used on flat surfaces while breaking no actual use case for what it replaces.