sprites: adjust PISGA0 perspective. (#1199)
The recent two-handed sprite update raised the sprite further up the screen - which looks good, except the gun now looks like it's pointing somewhere above the center line of the screen.

This flattens the A frame out a bit. B is left as-is as it just ends up looking like recoil when you shoot.