manual: Fix warning in tutorial steps 1-15 (#1225)
asciidoctor-pdf outputs a warning when ordered lists are numbered
entirely with "1.". It wants actual consecutive integers or ".". Prior
to this fix there were 14 warnings such as:

  asciidoctor: WARNING: manual.adoc: line 234: list item index: expected 2, got 1

in the English and French manual. Interestingly it makes no difference
in the PDF produced, but let's minimize the number of warnings in the
build.