levels: fix some E1M2 aesthetics. (#1247)
Closes #1245.

Fix that bridge to look a little less awkward and #996-compliant.

Add a decorative door to the starting area monster closet, implying the monsters are coming in from another, inaccessible part of the facility. It does, however, make the floor match the main area again.

Make the freestanding AGM logo bolder, simpler and stronger.

Adjust some textures in and near the big skylight hall to lean into the quasi-outdoors aesthetic while also replacing the grate floor textures with something a bit easier on the eyes in low-res.