Update dehacked.txt
Okay, this one's a very long one.


MAP06: Was changed in 0.12 to a new map.
MAP09: Space out "data" and "center".
MAP23: Have it go with the rest of Cluster 3 with this commit.
MAP24: Was changed in 0.11 to a new map - reusing mid-development MAP11's name.
MAP25: Changed because this name was somewhat based off of "Bloodfalls" from Doom II, according to the person who named it. It's quite bland as well. - reusing mid-development E2M1's name
MAP29: "Last Stand" is very bland and sticks out like a sore thumb against the rest of Cluster 3.