dehacked: rename levels as agreed in #1264 + E3M2/E3M7. (#1267)
E1M7: There was some talk about reverting to "Viral Research" but the viral waste being kept in open air pits seemed no more plausible than this building being a "transportation bay". "Materials" was evocative of something that would probably stay put if left alone. (Probably.)

E3M2: There never was anything added to this level that actually resembled a temple, though there are plenty of tunnels and things evocative of heat. There's also machinery suggesting some kind of power generation, and "Heate Tunnels" seemed oddly bathetic and understandably replaced.

DM22: Fixed inconsistent capitalization.

Map02: This level didn't really resemble a lab, let alone multiple of them. It was originally named "Jade Tumult" but that name doesn't fit this cluster at all; another name with the same initials was suggested, with "Janus" for the two-faced appearance of the starting area and "Terminal" for the big end teleporter (and cluster finale).

Map24: This is a dark, tight, claustrophobic map with narrow doorways and a very uneven floor, and in no way resembles something that can be called a "loading bay". It does, however, contain an odd arrangement of teleporters that does not appear to serve any practical purpose.

Map25: "Red Works" seemed still too reminiscent of id "Bloodfalls". "Surreality" was initially suggested by Uni but I disagreed because I always associated that name with the old E2M1, so we agreed to split the difference by naming it after a famous surrealist painting that involved melting and impossible machines.

Map29: "Last Stand" is not a stand, nor last. There's nothing about the flow of the level that would suggest anyone's doing a "last stand" here. The most prominent thing about it is the vast expanse of lava that some of the architecture is melting into. "Lake of Fire" was rejected as too doomily biblical.