Flip E2M3 and DM06's tracks (#1271)
"Tunnel" feels out of place among E2's tracks and a lot more suited for the sandy, Egyptian temple-styled DM06. The feeling is coming from the harpsichord and thrashy instrumentation.

"Workout" is a lot more upbeat and fitting for the first few E2 tracks, Korp had originally intended it for E2M1 also.