levels: use old slime bricks for Map11 west entry.
This is actually a perfect example of why I wanted to keep the old slime bricks in #1229. This is supposed to be a very poorly maintained, badly worn-out, poorly lit, barely structurally functional section of the building that the new bricks just look far too solid and *intact* to convey properly.