levels: fix minor E2M3 lighting change issue.
After the lights come back on in the blue room the lights in there appear to cast darkness instead of light. A bunch of stuff has been rearranged slightly so that the brightest adjacent sectors to the affected ones are not brighter than these light sectors.

As a side effect, this means that the gradient of the big techpillar light by the entrance is less affected by the in-game shutoff than it used to be. Unless the player has enough minigun ammo to camp there on purpose I don't think it'll affect gameplay for most people.

I've also simplified that room to use much fewer sectors - none of the lines between them were sound blockers and there were no changes in floor height or texture or lighting.