dist: Use non-thumbnails for screenshots. (#1333)
Switch to using the full screenshots in the metainfo files. These URLs should not change.

I've checked them, and I've added a comment that should help with the confusion as a result of their being both PNGs and JPGs. My impression is these should be pulled infrequently such as when flatpak and Linux repo packaging is done each Freedoom release, or Linux distro release.