levels: improve Map29 minigunner trap.
The door switch is now slightly better lit, and is on the bright side of the fake contrast, while the teleport switch is hidden in the door-side corner on the dark side of the fake contrast. They're both visible objects at vanilla eye level. Textures are adjusted so it's a gradual fade of the bloodrust towards the cleanness of the switches rather than an obvious texture mismatch.

The doors are fully repeatable and fast when opened from outside, and their tops are lowered just a little bit (mostly to match the wood texture) - you should be noticing the minigunners just in time for the doors to close.

The minigunners are marked as ambush for good measure.