levels: misc. Map08 fixes.
Fixed the mismatched lighting under the switch room rising pillars.

Simplified some of the light gradients to help reduce some visplanes. The ledge on the east side of the horseshoe is no longer lit.

The newly lit door next to the ledge now has a similar light texture trim to the other lit areas.