New E1M1
While the old E1M1 is fun, I feel it's a bit too much at once, and in no particular order, not to mention leaving the rest of the episode without much to surprise the player or shake things up. I find it to not be a very good introductory map, and so I made this. This map is simpler, with secrets to unlock your arsenal for players that look beyond those corridors on the surface. But the enemy variety is stripped back slightly, and introduced at a slower pace, that way pinkies, specters, and shotgunners later introduced can serve to keep the gameplay fresh throughout.

The old E1M1 is not gone. I found the level rather fun, and so I have created an expanded version, not found here, but to be potentially added later on once polished enough for me to put it up, hopefully to be used somewhere around the mid point of E1.