levels: fix some visplane errors. (#1382)
e2m2
Visplane Explorer blob of red in the northeast by the little screens in Visplane Explorer when all 3 doors in the final area are pre-opened. Simplified some of the lighting and also got rid of the upper floor trim in favour of using a flat there that was lined in a way that lined up with the big walls.

e2m9
It was almost immediate the moment you opened the door on sector 385 from the east, not sure how we didn't catch that a lot earlier.
The main culprit - again - is the trim around the hex floor. I've gotten rid of that, but have not simplified or adjusted anything otherwise.