levels: fix E4M8 final swarm + water area secret.
The detailing on the way the airlock of the ship docks to the starport had one spot that was too steep for monsters to move over, trapping them inside.

The secret switch behind the pressure plate in the water is impossible to see in vanilla. The switch is now the skull wall itself.