textures: add flipped versions of AQMETL16-17.
Both of these chain textures should be usable as decorations in places the player can pass through, but are not complicated enough that the fact that they don't look the same from either side would get lost in the details. I don't see why they shouldn't deserve similar treatment to the diagonal chains (18 and 19).