Sounds: Punching
Back when I and a couple other submited the Octaminator, I included a new punching sound for it in the package. I wanted to make it's punch sound stronger than the player's punch (like it is in Doom with the revenant) so I gave it some more bass and lowered the pitch a tiny bit, but in the end it just sounds kinda muddy.

To solve this I went the opposite direction this time. I gave the current more over-the-top punching sound we have as is (unedited) to the Octaminator, and for the player I edited it to be a bit shorter and more subdued, but still satisfying.