Simplify the help screen
This fixes #1489, where the keys were being obscured by the menu skull
that is drawn on the help screen by vanilla Doom. Rather than just
shuffle around the contents of the help screen, I opted to simplify it
by removing a lot of the clutter.

The old help screen was created before the Freedoom manual was added, so
there's no need for us to include every in-game item. Instead, we can
delegate to the manual by including a link to it, and cut things back to
just a small selection of example items.