manual: Initial version of manual cover
Having this will allow us to put up printed Freedoom manuals for sale.
The SVG here is compatible with the template provided by lulu.com, which
is where I'm planning to put up the online store. Lulu expects a PDF, so
there's a Makefile target that uses Inkscape to convert the SVG to PDF
format.

The cover artwork here is by Reddit user Illustrious-Gift4206 who
created fanart of many of the Freedoom monsters. Ony my request he has
graciously agreed to donate his artwork to the project for use.

Currently this is just the English version of the manual cover. I'm
planning in a follow-up commit to convert the SVG into a template, so
that we can autogenerate covers for all the different translations.