manual: Add cover PDF build to CI
This is not invoked as part of the normal build process, and shouldn't
be: we only ever want to build the cover PDFs when we're uploading them
for printing, and don't want everyone who wants to build Freedoom to
have to install Inkscape. However, we should nonetheless include it in
the CI to check that the build step works.