manual: Fix problems uploading PDF to Lulu.com.
When uploading to Lulu.com an error was being produced, and with some
experimentation I discovered it was due to the `[%always]` tag used to
force page breaks; removing it fixes the problem. To ensure the Table of
Contents appears on the right page, I added a small amount of copyright
blurb to page 2.