README: Deemphasize compatibility
This moves the paragraph about compatibility further down the page into
its own section. I am *not* doing this because we are deemphasizing
compatibility as a project goal; rather, I'm doing this because I want
to *emphasize* the idea of Freedoom being its own independent game. The
idea here is to try to get away from the misconception some people have
that the goal of Freedoom is to be a "clone" of Doom.