README: Reword intro section
The previous wording ("but Freedoom by itself is just...") read like we
were putting the project down a bit. Let's not do ourselves a
disservice.

A second paragraph further down was framing us in terms of being a "free
alternative" to Doom. Let's ditch that; we are our own game. We can also
leave out the technical jargon about IWADs etc.