README: Make IP advice more explicit/forceful
The previous text was couching its words a lot ("don't emulate...
exactly", "this is a tough call"). Let's be more explicit and emphasize
that work should be original and not a clone of Doom. Besides being
safer from a legal standpoint, it's a better goal for the project in
general.