levels: fix DM24 and DM25 vanilla issues.
Just a couple things I accidentally found tonight.

DM24: In the northwest, if you hit the light grey pillars (the ones with the corners filed off) at just the right angle, the projectiles would disappear. Turns out a huge part of that upper trim's sky-facing linedefs were facing the wrong way.

DM25: The use of BIGDOOR6 for those steps was causing tutti-frutti. This is exactly the sort of thing BIGDOR6A was made for - all 21 occurrences of this texture replaced accordingly.