Revert "Updated Charli's E1M1 map. (#1496)"
This reverts commit 9b25e1447aac8bfcab8a36815a32bb73d27263ce.

The first level of the game is very prominent and I'm not convinced that
this had sufficient debate on the original PR. Opinions on the original
PR seemed mixed to lukewarm.

I'm also not convinced that this is an improvement over the previous
E1M1 but that is just my personal opinion. If we want to make this
change then we can, but it should be discussed more first.