sprites: improve shotgun guy death.
This has bothered me for years but it goes by so quick you never really notice it: the shotgun guy's entire abdomen and pelvic area are missing in multiple frames. The body is now restored.

The blood on the corpse is also made a little more consistent, and treads have been added to the soles of the boots.